# services/faux-schoolday/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Faux Schoolday API
  description: |
    High-fidelity mock implementation of Schoolday platform APIs for teacher evaluation system integration.

    This mock service provides comprehensive simulation of:
    - OAuth 2.0 authentication flows
    - OneRoster v1.1 standard for educational data exchange
    - Discovery API for district and school information
    - Academy API for professional development
    - Webhook system for real-time event notifications

    All APIs are designed to match Schoolday's actual behavior for seamless local development and testing.
  version: 1.0.0
  contact:
    name: Teacher Evaluation Platform Team
    email: dev@teacher-eval.local
  license:
    name: MIT License
servers:
  - url: http://localhost:3006/api
    description: Local development server
  - url: https://faux-schoolday.teacher-eval.local/api
    description: Containerized environment
security:
  - BearerAuth: []
paths:
...
  /webhooks/subscriptions:
    get:
      tags: [Webhooks]
      summary: List webhook subscriptions
      parameters:
        - name: organization_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, disabled]
      responses:
        '200':
          description: List of webhook subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionListResponse'

    post:
      tags: [Webhooks]
      summary: Create webhook subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookSubscriptionRequest'
      responses:
        '201':
          description: Webhook subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WebhookSubscription:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        organization_id:
          type: string
          nullable: true
        description:
          type: string
        status:
          type: string
          enum: [active, paused, disabled]
        created_at:
          type: string
          format: date-time
        last_delivery:
          type: string
          format: date-time
          nullable: true
        delivery_success_rate:
          type: number
        total_deliveries:
          type: integer
        failed_deliveries:
          type: integer
    WebhookSubscriptionListResponse:
      type: object
      properties:
        subscriptions:
          type: array
          items:
            $ref: '#/components/schemas/WebhookSubscription'
        total:
          type: integer
    CreateWebhookSubscriptionRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        organization_id:
          type: string
          nullable: true
        description:
          type: string
        secret:
          type: string
      required: [url, events]
    WebhookSubscriptionResponse:
      type: object
      properties:
        subscription:
          $ref: '#/components/schemas/WebhookSubscription'

    # Common Schemas
...